refactor(chat): remove duplicate transcript block projection
