--- src/renderer/src/components/workspace/chat/timelineBlocks.ts
@@ -1,6 +1,0 @@
-export {
-    TimelineBlockCache,
-    type TimelineBlock,
-    type TimelineBlockPreferences,
-    type TimelineBlockRow,
-} from "@renderer/app/ai/timelineBlockCache";
